[Mobile App] Add support for standalone app
Currently when we use the debug build the application needs to have the
npm server running in the developer PC. The required JS resources needed
to run the application are downloaded on first run over ADB. Add the
required bundle to the app resource so that it can work without the node
server.

- Change the build.gradle to solve the problem when building the
standalone application
- Updated the README with the steps to generate the standalone
application.

Change-Id: Id6a56f5feca5b6221f3b57bd6d0e988d9f596738
Signed-off-by: Nisarg Parmar <nisarg.parmar@einfochips.com>